+++ Tests/GRDBTests/ValueObservationTests.swift
@@ -523,6 +523,49 @@ class ValueObservationTests: GRDBTestCase {
     }
 #endif
     
+    // MARK: - Unspecified Changes
+    
+    func test_ValueObservation_is_triggered_by_explicit_change_notification() throws {
+        let dbQueue1 = try makeDatabaseQueue(filename: "test.sqlite")
+        try dbQueue1.write { db in
+            try db.execute(sql: "CREATE TABLE test(a)")
+        }
+        
+        let undetectedExpectation = expectation(description: "undetected")
+        undetectedExpectation.expectedFulfillmentCount = 2 // initial value and change
+        undetectedExpectation.isInverted = true
+
+        let detectedExpectation = expectation(description: "detected")
+        detectedExpectation.expectedFulfillmentCount = 2 // initial value and change
+        
+        let observation = ValueObservation.tracking { db in
+            try Table("test").fetchCount(db)
+        }
+        let cancellable = observation.start(
+            in: dbQueue1,
+            scheduling: .immediate,
+            onError: { error in XCTFail("Unexpected error: \(error)") },
+            onChange: { _ in
+                undetectedExpectation.fulfill()
+                detectedExpectation.fulfill()
+            })
+
+        try withExtendedLifetime(cancellable) {
+            // Change performed from external connection is not detected...
+            let dbQueue2 = try makeDatabaseQueue(filename: "test.sqlite")
+            try dbQueue2.write { db in
+                try db.execute(sql: "INSERT INTO test (a) VALUES (1)")
+            }
+            wait(for: [undetectedExpectation], timeout: 2)
+            
+            // ... until we perform an explicit change notification
+            try dbQueue1.write { db in
+                try db.notifyChanges(in: Table("test"))
+            }
+            wait(for: [detectedExpectation], timeout: 2)
+        }
+    }
+    
     // MARK: - Cancellation
     
     func testCancellableLifetime() throws {